Move impl ReturnShipping for *Iterator into pgrx::iter
purely administrative change
